Release rk3308 - 2023-12-29
Based on:
    e237b9d configs: debian: Add Debian12 (Bookworm) to support

+++ common/scripts/post-async-commit.sh
@@ -2,15 +2,17 @@
 
 POST_ROOTFS_ONLY=1
 
-source "${POST_HELPER:-$(dirname "$(realpath "$0")")/../post-hooks/post-helper}"
+source "${RK_POST_HELPER:-$(dirname "$(realpath "$0")")/../post-hooks/post-helper}"
 
-echo "Installing async-commit service..."
+[ "$RK_ROOTFS_ASYNC_COMMIT" ] || exit 0
+
+message "Installing async-commit service..."
 
 rm -f etc/init.d/S*_commit.sh \
 	etc/systemd/system/multi-user.target.wants/async.service \
 	usr/lib/systemd/system/async.service
 
-cd "$SDK_DIR"
+cd "$RK_SDK_DIR"
 
 mkdir -p "$TARGET_DIR/usr/bin"
 install -m 0755 external/rkscript/async-commit "$TARGET_DIR/usr/bin/"
